Added rash computation algo
