--- core/banabout.py
@@ -1,29 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-
-#-:-:-:-:-:-:-::-:-:#
-#    XSRF Probe     #
-#-:-:-:-:-:-:-::-:-:#
-
-#Author: @_tID
-#This module requires XSRFProbe
-#https://github.com/0xInfection/XSRFProbe
-
-import time
-from core.colors import *
-
-def banabout(): # some fancy banner stuff :p
-
-    print(color.BLUE+'   [---]           '+color.GREY+'XSRF Probe |'+color.RED+' A'+color.ORANGE+' Cross Site Request Forgery '+color.RED+'Audit Toolkit          '+color.BLUE+'[---]')
-    time.sleep(0.2)
-    print(color.BLUE+'   [---]                                                                            [---]')
-    time.sleep(0.2)
-    print(color.BLUE+'   [---]   '+color.PURPLE+'                  '+color.GREEN+'~  Author : '+color.CYAN+'The Infected Drake  ~                 '+color.BLUE+'     [---]')
-    time.sleep(0.2)
-    print(color.BLUE+'   [---]   '+color.CYAN+'                    ~  github.com / '+color.GREY+'0xInfection  ~                     '+color.BLUE+'  [---]')
-    time.sleep(0.2)
-    print(color.BLUE+'   [---]                                                                            [---]')
-    time.sleep(0.2)
-    print(color.BLUE+'   [---]   '+color.ORANGE+'                          ~  Version '+color.RED+'v1.1.0'+color.ORANGE+'  ~                         '+color.BLUE+'  [---]\n')
-    time.sleep(0.2)
-
